requirements.txt
httpx~=0.28.1
pydantic~=2.10.4
PyJWT==2.9.0 # Pinned because of Python 3.8 incompatibility in future versions
cryptography~=44.0.0
